Move expanding/collapsing logic into SolutionNodesExpansionState.
